+++ tests/utils/webhook.py
@@ -6,7 +6,7 @@ def generateExampleWebhook():
     return starkinfra.Webhook(
         url="https://webhook.site/{uuid}".format(uuid=str(uuid4())),
         subscriptions=[
-            "credit-note", "signer",
+            "credit-note",
             "issuing-card", "issuing-invoice", "issuing-purchase",
             "pix-request.in", "pix-request.out", "pix-reversal.in", "pix-reversal.out", "pix-claim", "pix-key", "pix-chargeback", "pix-infraction",
         ],