+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingBin resource to IssuingProduct
 - fine and interest attributes to return only on CreditNote.Invoice sub-resource
 - expiration in the CreditNote resource from returned-only to optional parameter
+- Creditnote.Signer sub-resource to CreditSigner resource
 
 ## [0.1.1] - 2022-06-09
 ### Fixed